+++ Font-Awesome/Font-Awesome/src/main/scala/net/liftmodules/FoBoFA/FoBoFA.scala
@@ -79,22 +79,21 @@ package object FoBoFA {
     override def toString() = "FoBoFA.ToolKit = "+store.toString()
     
    /**
-     * Enable usage of FoBo's FontAwesome resources and API version 4&#8228;5&#8228;0 in your bootstrap liftweb Boot.
-     * @version 4.5.0
+     * Enable usage of FoBo's FontAwesome resources and API version 4&#8228;6&#8228;3 in your bootstrap liftweb Boot.
+     * @version 4.6.3
      * 
      * '''Example:'''
      * 
      * {{{
      *   import net.liftmodules.{FoBoFA => FoBo}
      *    :
-     *   FoBo.ToolKit.Init=FoBo.ToolKit.FontAwesome450
+     *   FoBo.ToolKit.Init=FoBo.ToolKit.FontAwesome463
      * }}}
      */    
-     case object FontAwesome450 extends ToolKit {
-       FoBoFARes.Resource.FontAwesome450
+     case object FontAwesome463 extends ToolKit {
+       FoBoFARes.Resource.FontAwesome463
        //FoBoFAAPI.API.FontAwesome4
      } 
-    
     
     /**
      * Enable usage of FoBo's FontAwesome resources and API version 4&#8228;3&#8228;0 in your bootstrap liftweb Boot.
@@ -181,20 +180,21 @@ package object FoBoFA {
       store
     }     
     override def toString() = "FoBoFA.Resource = "+store.toString()
+    
     /**
-     * Enable usage of FoBo's FontAwesome resources version 4&#8228;5&#8228;0 in your bootstrap liftweb Boot.
-     * @version 4.5.0
+     * Enable usage of FoBo's FontAwesome resources version 4&#8228;6&#8228;3 in your bootstrap liftweb Boot.
+     * @version 4.6.3
      * 
      * '''Example:'''
      * 
      * {{{
      *   import net.liftmodules.{FoBoFA => FoBo}
      *    :
-     *   FoBo.Resource.Init=FoBo.Resource.FontAwesome450
+     *   FoBo.Resource.Init=FoBo.Resource.FontAwesome463
      * }}}
      */    
-     case object FontAwesome450 extends Resource {
-       FoBoFARes.Resource.FontAwesome450
+     case object FontAwesome463 extends Resource {
+       FoBoFARes.Resource.FontAwesome463
      }     
     
     /**